pyproject.toml
[project]
name = "unified-retreival-engine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=26.1.0",
    "datasets>=4.5.0",
    "docling>=2.68.0",
    "docling-core>=2.59.0",
    "faker>=40.1.2",
    "fastapi[standard]>=0.128.0",
    "httpx>=0.28.1",
    "hydra-core>=1.3.2",
    "numpy>=2.4.1",
    "openai>=2.15.0",
    "opik>=1.9.87",
    "pdfminer-six==20231228",
    "pillow>=11.3.0",
    "pinecone[grpc]>=8.0.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.1",
    "python-magic>=0.4.27",
    "python-multipart>=0.0.21",
    "ragas>=0.4.3",
    "ruff>=0.14.13",
    "semchunk>=2.2.2",
    "sqlalchemy>=2.0.45",
    "tiktoken>=0.12.0",
    "unstructured[all-docs]>=0.18.27",
    "vanna[fastapi,openai,postgres]>=2.0.1",
]
